create new communcation to converting service+Dockerfiles+QuoteClient for Converting+logic to convert file

+++ day-service/src/main/java/com/quoteForDay/dayservice/model/Day.java
@@ -12,11 +12,11 @@ public class Day {
 
         @Id
         @GeneratedValue(strategy = IDENTITY)
-        private Long id;
+        private int id;
 
         private String day_desc;
 
-        Day(Long id, String day_desc) {
+        Day(int id, String day_desc) {
             this.id=id;
             this.day_desc=day_desc;
         }
